Updated compatibility.
When searching for best option to get all FilterRules for jQuery QueryBuilder, I cannot find this NuGet package nor cannot find it on Google. Upon further search, the problem is that jQuery QueryBuilder on the description has a dash on it, so I removed it and added a link so that this tool will show on google search results when someone will look for dynamic linq query builder just like I did.