"In Query" is not working.
This is an update on PR related to this previous request:

https://github.com/castle-it/dynamic-linq-query-builder/pull/31

As suggested by @tghamm, the tolower comparisons for strings are handled in each individual operator type. So I re-adjust in on static Expression "In" method.

I discovered the issue when I am querying a string with all uppercase. It always return zero results.